# Optimize empty patterns case