@andrewdalpino 2020-09-26
Hey @simplechris I don't think we should be too worried about the performance in this case since its occurence will be rare and usually special (either the user temporarily cleared their `$stopWords` variable or didn't know what they were doing. The point of those changes were to not throw an exception when we don't have to. Would you say this is premature optimization? That is, optimizing an edge case?